Make lokapi into a separate service file
