+++ src/main.ts
@@ -2,7 +2,7 @@ import { createApp } from "vue";
 import App from "./App.vue";
 import router from "./router";
 import store from "./store";
-import { lokAPI } from "./store/lokapi";
+import { lokApiService } from "./services/lokapiService"
 import Toaster from "@meforma/vue-toaster";
 import Swal from "./useSwal";
 require("@/assets/main.scss");
@@ -17,5 +17,5 @@ const app = createApp(App)
   app.use(Swal)
   app.use(Toaster)
   app.provide('$store', store)
-  app.config.globalProperties.$lokapi = lokAPI
+  app.config.globalProperties.$lokapi = lokApiService
   app.mount("#app");
